# project-api/out/swagger_server/models/new_training.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class NewTraining(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, type: str=None, configuration: str=None, artifacts: List[str]=None, reports: List[str]=None, creation_time: datetime=None, last_update_time: datetime=None):  # noqa: E501
        """NewTraining - a model defined in Swagger

        :param type: The type of this NewTraining.  # noqa: E501
        :type type: str
        :param configuration: The configuration of this NewTraining.  # noqa: E501
        :type configuration: str
        :param artifacts: The artifacts of this NewTraining.  # noqa: E501
        :type artifacts: List[str]
        :param reports: The reports of this NewTraining.  # noqa: E501
        :type reports: List[str]
        :param creation_time: The creation_time of this NewTraining.  # noqa: E501
        :type creation_time: datetime
        :param last_update_time: The last_update_time of this NewTraining.  # noqa: E501
        :type last_update_time: datetime
        """
        self.swagger_types = {
            'type': str,
            'configuration': str,
            'artifacts': List[str],
            'reports': List[str],
            'creation_time': datetime,
            'last_update_time': datetime
        }

        self.attribute_map = {
            'type': 'type',
            'configuration': 'configuration',
            'artifacts': 'artifacts',
            'reports': 'reports',
            'creation_time': 'creation_time',
            'last_update_time': 'last_update_time'
        }
        self._type = type
        self._configuration = configuration
        self._artifacts = artifacts
        self._reports = reports
        self._creation_time = creation_time
        self._last_update_time = last_update_time

    @classmethod
    def from_dict(cls, dikt) -> 'NewTraining':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The NewTraining of this NewTraining.  # noqa: E501
        :rtype: NewTraining
        """
        return util.deserialize_model(dikt, cls)

    @property
    def type(self) -> str:
        """Gets the type of this NewTraining.


        :return: The type of this NewTraining.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this NewTraining.


        :param type: The type of this NewTraining.
        :type type: str
        """
        if type is None:
            raise ValueError("Invalid value for `type`, must not be `None`")  # noqa: E501

        self._type = type

    @property
    def configuration(self) -> str:
        """Gets the configuration of this NewTraining.


        :return: The configuration of this NewTraining.
        :rtype: str
        """
        return self._configuration

    @configuration.setter
    def configuration(self, configuration: str):
        """Sets the configuration of this NewTraining.


        :param configuration: The configuration of this NewTraining.
        :type configuration: str
        """

        self._configuration = configuration

    @property
    def artifacts(self) -> List[str]:
        """Gets the artifacts of this NewTraining.


        :return: The artifacts of this NewTraining.
        :rtype: List[str]
        """
        return self._artifacts

    @artifacts.setter
    def artifacts(self, artifacts: List[str]):
        """Sets the artifacts of this NewTraining.


        :param artifacts: The artifacts of this NewTraining.
        :type artifacts: List[str]
        """

        self._artifacts = artifacts

    @property
    def reports(self) -> List[str]:
        """Gets the reports of this NewTraining.


        :return: The reports of this NewTraining.
        :rtype: List[str]
        """
        return self._reports

    @reports.setter
    def reports(self, reports: List[str]):
        """Sets the reports of this NewTraining.


        :param reports: The reports of this NewTraining.
        :type reports: List[str]
        """

        self._reports = reports

    @property
    def creation_time(self) -> datetime:
        """Gets the creation_time of this NewTraining.


        :return: The creation_time of this NewTraining.
        :rtype: datetime
        """
        return self._creation_time

    @creation_time.setter
    def creation_time(self, creation_time: datetime):
        """Sets the creation_time of this NewTraining.


        :param creation_time: The creation_time of this NewTraining.
        :type creation_time: datetime
        """

        self._creation_time = creation_time

    @property
    def last_update_time(self) -> datetime:
        """Gets the last_update_time of this NewTraining.


        :return: The last_update_time of this NewTraining.
        :rtype: datetime
        """
        return self._last_update_time

    @last_update_time.setter
    def last_update_time(self, last_update_time: datetime):
        """Sets the last_update_time of this NewTraining.


        :param last_update_time: The last_update_time of this NewTraining.
        :type last_update_time: datetime
        """

        self._last_update_time = last_update_time