# project-api/out/swagger_server/models/runtime.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.job import Job  # noqa: F401,E501
from swagger_server import util


class Runtime(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, type: str=None, jobs: List[Job]=None):  # noqa: E501
        """Runtime - a model defined in Swagger

        :param type: The type of this Runtime.  # noqa: E501
        :type type: str
        :param jobs: The jobs of this Runtime.  # noqa: E501
        :type jobs: List[Job]
        """
        self.swagger_types = {
            'type': str,
            'jobs': List[Job]
        }

        self.attribute_map = {
            'type': 'type',
            'jobs': 'jobs'
        }
        self._type = type
        self._jobs = jobs

    @classmethod
    def from_dict(cls, dikt) -> 'Runtime':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Runtime of this Runtime.  # noqa: E501
        :rtype: Runtime
        """
        return util.deserialize_model(dikt, cls)

    @property
    def type(self) -> str:
        """Gets the type of this Runtime.


        :return: The type of this Runtime.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this Runtime.


        :param type: The type of this Runtime.
        :type type: str
        """

        self._type = type

    @property
    def jobs(self) -> List[Job]:
        """Gets the jobs of this Runtime.


        :return: The jobs of this Runtime.
        :rtype: List[Job]
        """
        return self._jobs

    @jobs.setter
    def jobs(self, jobs: List[Job]):
        """Sets the jobs of this Runtime.


        :param jobs: The jobs of this Runtime.
        :type jobs: List[Job]
        """

        self._jobs = jobs