# project-api/out/swagger_server/swagger/swagger.yaml
openapi: 3.0.2
info:
  title: STAIoTCraft - Project API
  description: REST API to access STAIoTCraft Back-End web-service User Projects
  version: 3.0.5
servers:
- url: /3
paths:
...
  /projects/{project_name}/models/{model_name}/{activity_type}:
    summary: Activity data resource
    description: ""
    get:
      tags:
      - Activity
      summary: "Get the file contents of artifacts, reports or configuration of the\
        \ activity"
      description: ""
      operationId: app_get_activity
      parameters:
      - name: project_name
        in: path
        description: Project `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: model_name
        in: path
        description: Model `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: activity_type
        in: path
        description: Activity `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: type
        in: query
        description: "string corresponding to artifact to be GET artifacts, reports,\
          \ runtime OR config"
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: name
        in: query
        description: Filename of artifact for which GET is issued
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                x-content-type: application/json
        "400":
          description: Bad request
        "401":
          description: Unathorized
        "404":
          description: Not found
      security:
      - bearerAuth:
        - secret
      x-openapi-router-controller: project_api.controllers.activity_controller
    put:
      tags:
      - Activity
      summary: Update Activity JSON
      description: ""
      operationId: app_update_activity
      parameters:
      - name: project_name
        in: path
        description: Project `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: model_name
        in: path
        description: Model `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: activity_type
        in: path
        description: Activity `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: The activity to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Training'
        required: true
      responses:
        "201":
          description: Created.
        "400":
          description: Bad request.
        "401":
          description: Unauthorized.
        "404":
          description: Not found.
      security:
      - bearerAuth:
        - secret
      x-openapi-router-controller: project_api.controllers.activity_controller
    post:
      tags:
      - Activity
      summary: Create new Activity
      description: ""
      operationId: app_create_activity
      parameters:
      - name: project_name
        in: path
        description: Project `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: model_name
        in: path
        description: Model `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: activity_type
        in: path
        description: Activity `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: The activity to be added.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTraining'
        required: true
      responses:
        "201":
          description: Created.
        "400":
          description: Bad request.
        "401":
          description: Unauthorized.
        "404":
          description: Not found.
      security:
      - bearerAuth:
        - secret
      x-openapi-router-controller: project_api.controllers.activity_controller
    delete:
      tags:
      - Activity
      summary: Delete the activity associated to the given name.
      description: ""
      operationId: app_delete_activity
      parameters:
      - name: project_name
        in: path
        description: Project `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: model_name
        in: path
        description: Model `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: activity_type
        in: path
        description: Activity `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "400":
          description: Bad request
        "401":
          description: Unathorized
        "404":
          description: Not found
      security:
      - bearerAuth:
        - secret
      x-openapi-router-controller: project_api.controllers.activity_controller
    patch:
      tags:
      - Activity
      summary: Update partial Activity JSON
      description: ""
      operationId: app_patch_activity
      parameters:
      - name: project_name
        in: path
        description: Project `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: model_name
        in: path
        description: Model `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: activity_type
        in: path
        description: Activity `name` identifier
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: The activity to be patched.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/model_name_activity_type_body'
        required: true
      responses:
        "201":
          description: Created.
        "400":
          description: Bad request.
        "401":
          description: Unauthorized.
        "404":
          description: Not found.
      security:
      - bearerAuth:
        - secret
      x-openapi-router-controller: project_api.controllers.activity_controller
...
components:
  schemas:
...
    NewTraining:
      required:
      - type
      type: object
      properties:
        type:
          type: string
        configuration:
          type: string
        artifacts:
          type: array
          items:
            type: string
        reports:
          type: array
          items:
            type: string
        creation_time:
          type: string
          format: date-time
        last_update_time:
          type: string
          format: date-time
      example:
        type: job
        configuration: MLC_Configuration.json
        artifacts:
        - artifacts
        - artifacts
        reports:
        - reports
        - reports
        creation_time:
        - creation_time
        last_update_time:
        - last_update_time
...
    Training:
      required:
      - artifacts
      - configuration
      - reports
      - runtime
      type: object
      properties:
        runtime:
          $ref: '#/components/schemas/Runtime'
        configuration:
          type: string
        artifacts:
          type: array
          items:
            type: string
        reports:
          type: array
          items:
            type: string
        creation_time:
          type: string
          format: date-time
        last_update_time:
          type: string
          format: date-time
      example:
        runtime:
          job_id: job_id
          tool: tool
          version: version
        configuration: configuration.json
        artifacts:
        - artifacts
        - artifacts
        reports:
        - reports
        - reports
        creation_time:
        - creation_time
        last_update_time:
        - last_update_time
    Runtime:
      type: object
      properties:
        type:
          type: string
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
      example:
        type: job
        jobs:
        - jobs
        - jobs
    Job:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        template_id:
          type: string
      example:
        name: job-alias
        version: job-version
        template_id: job-template-id
...
    model_name_activity_type_body:
      type: object
      properties:
        configuration:
          type: string
        artifacts:
          type: array
          items:
            type: string
        reports:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth